Extract (and rename) DbtCloudTransformationsCard
