+++ airbyte-webapp/src/pages/ConnectionPage/pages/ConnectionItemPage/ConnectionTransformationTab.tsx
@@ -1,12 +1,8 @@
-import { faPlus } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import classNames from "classnames";
 import { Field, FieldArray } from "formik";
 import React, { useMemo } from "react";
 import { FormattedMessage } from "react-intl";
 import { useToggle } from "react-use";
 
-import { Button } from "components/ui/Button";
 import { Card } from "components/ui/Card";
 import { Text } from "components/ui/Text";
 
@@ -28,7 +24,7 @@ import {
 } from "views/Connection/ConnectionForm/formConfig";
 import { FormCard } from "views/Connection/FormCard";
 
-import styles from "./ConnectionTransformationTab.module.scss";
+import { DbtCloudTransformationsCard } from "./ConnectionTransformationTab/DbtCloudTransformationsCard";
 
 const CustomTransformationsCard: React.FC<{
   operations?: OperationCreate[];
@@ -66,40 +62,6 @@ const CustomTransformationsCard: React.FC<{
         )}
       </FieldArray>
     </FormCard>
-  );
-};
-
-const CloudTransformationsCard = () => {
-  // TODO fetch list of transformations for real
-  const transformations = [];
-  /* const transformations = [1, 2, 3]; */
-
-  const TransformationList = ({ className }: { className: string }) =>
-    transformations.length ? (
-      <div className={className}>this is a list</div>
-    ) : (
-      <div className={classNames(className, styles.emptyListContent)}>
-        <div className={styles.contextExplanation}>
-          After an Airbyte sync job has completed, the following jobs will run.
-        </div>
-        <img src="/images/octavia/worker.png" alt="An octopus wearing a hard hat, tools at the ready" />
-        No transformations
-      </div>
-    );
-
-  return (
-    <Card
-      title={
-        <span className={styles.cloudTransformationsListTitle}>
-          Transformations
-          <Button variant="secondary" icon={<FontAwesomeIcon icon={faPlus} />}>
-            Add transformation
-          </Button>
-        </span>
-      }
-    >
-      <TransformationList className={styles.cloudTransformationsListContainer} />
-    </Card>
   );
 };
 
@@ -174,7 +136,7 @@ export const ConnectionTransformationTab: React.FC = () => {
       >
         {supportsNormalization && <NormalizationCard operations={connection.operations} onSubmit={onSubmit} />}
         {supportsDbt && <CustomTransformationsCard operations={connection.operations} onSubmit={onSubmit} />}
-        {supportsCloudDbtIntegration && <CloudTransformationsCard />}
+        {supportsCloudDbtIntegration && <DbtCloudTransformationsCard />}
         {noSupportedTransformations && (
           <Card className={styles.customCard}>
             <Text as="p" size="lg" centered>
